[1/2] Move HCAFile factory from repository to metadata plugin (#7496)
Move methods
